+++ backend/src/routes/routes.ts
@@ -3,12 +3,10 @@ import { Router } from 'express';
 import tasksRouter from './tasks.routes';
 import settingsRouter from './settings.routes';
 
-
 const routes = Router();
 
 routes.use('/tasks', tasksRouter);
 
 routes.use('/settings', settingsRouter);
-
 
 export default routes;
